move typings into root
