+++ src/init.ts
@@ -1,10 +1,10 @@
 import { Store } from 'redux'
+import { Config, ConfigRedux, Exposed } from '../typings/rematch'
 import { postStore, preStore } from './core'
 import { initModelHooks } from './model'
 import corePlugins from './plugins'
 import { initReducers } from './redux/reducers'
 import { initStore } from './redux/store'
-import { Config, ConfigRedux, Exposed } from './typings/rematch'
 import buildPlugins from './utils/buildPlugins'
 import getExposed from './utils/getExposed'
 import getModels from './utils/getModels'
