+++ src/redux/store.ts
@@ -1,7 +1,7 @@
 /* eslint no-underscore-dangle: 0 */
 import { applyMiddleware, createStore as _createStore, Middleware, Reducer, Store, StoreCreator } from 'redux'
+import { Config, Model } from '../../typings/rematch'
 import { pluginMiddlewares } from '../core'
-import { Config, Model } from '../typings/rematch'
 import { composeEnhancers } from './devtools'
 import { createModelReducer, createRootReducer, mergeReducers } from './reducers'
