+++ src/utils/getModels.ts
@@ -1,4 +1,4 @@
-import { Model, Models } from '../typings/rematch'
+import { Model, Models } from '../../typings/rematch'
 
 export default (models: Models): Model[] => {
   return Object.keys(models).map((name: string) => ({
